최종프로젝트/requirements.txt
babel==2.14.0
gitpython==3.1.43
jinja2==3.1.3
markdown==3.6
markupsafe==2.1.5
pillow==9.4.0
pywavelets==1.4.1
pyyaml==6.0.1
send2trash==1.8.3
absl-py==2.1.0
accelerate==0.26.1
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.3
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
config==0.5.1
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.10.0
cymem==2.0.8
debugpy==1.6.7
decorator==4.4.2
decord==0.6.0
deface==1.5.0
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.16.0
distlib==0.3.8
easydict==1.9
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.4
fastapi==0.110.3
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.3.1
ftfy==6.2.0
gast==0.4.0
gitdb==4.0.11
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.16.0
gradio-client==0.8.1
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
identify==2.5.36
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.1.0
importlib-resources==6.4.0
iopath==0.1.10
ipykernel==6.29.3
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
joblib==1.4.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.4.9
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-packaging==0.12.3
jupyter-server==2.14.0
jupyter-server-fileid==0.9.2
jupyter-server-terminals==0.5.3
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.7
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.1
jupyterlab-widgets==3.0.10
kaggle==1.6.12
keras==2.13.1
kiwisolver==1.4.5
langcodes==3.4.0
language-data==1.2.0
lazy-loader==0.4
libclang==18.1.1
marisa-trie==1.1.0
markdown-it-py==3.0.0
matplotlib==3.7.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
nbclassic==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.6
notebook-shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.16.2
open3d==0.13.0
opencv-contrib-python==4.5.5.64
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
opendatasets==0.1.22
opt-einsum==3.3.0
orjson==3.10.1
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pip==23.3.1
pkgutil-resolve-name==1.3.10
platformdirs==4.2.1
plotly==5.21.0
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
proglog==0.1.10
prometheus-client==0.20.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pydeck==0.9.0b1
pydub==0.25.1
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pyyaml-env-tag==0.1
pyzmq==24.0.1
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
roboflow==1.1.28
rpds-py==0.18.0
rsa==4.9
ruff==0.4.2
safetensors==0.4.3
salesforce-lavis==1.0.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==0.4.12
tinycss2==1.3.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torcheval==0.0.7
torchvision==0.18.0
tornado==6.4
tqdm==4.64.1
traitlets==5.14.3
transformers==4.25.0
triton==2.3.0
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing==3.7.4.3
typing-extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.5
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.0
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
werkzeug==3.0.2
wheel==0.41.2
widgetsnbextension==4.0.10
wrapt==1.16.0
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.18.1
